ocr_web_app/requirements.txt
flask==3.1.0
pytesseract==0.3.13
pillow==11.1.0
werkzeug==3.1.3